deleteDialog added for deals

+++ src/app/modules/vendors/pages/deals/deals.component.ts
@@ -1,10 +1,17 @@
-import { Component, OnInit, ElementRef, ViewChild } from "@angular/core";
+import {
+  Component,
+  OnInit,
+  ElementRef,
+  ViewChild,
+  TemplateRef
+} from "@angular/core";
 import { FormGroup, FormControl, Validators } from "@angular/forms";
 import { FranchiseDealsService } from "../../services/franchiseDeals.service";
 import { AngularFireStorage } from "@angular/fire/storage";
 import { Observable } from "rxjs";
 import { finalize } from "rxjs/operators";
-import { IsButton, IsModalService } from '../../../../lib';
+import { IsButton, IsModalService } from "../../../../lib";
+import { IsToasterService } from "../../../../lib/toaster";
 import { dealModel } from "../../models/dealModel";
 import { FranchiseInfoService } from "../../services/franchiseInfo.service";
 
@@ -25,8 +32,9 @@ export class DealsComponent implements OnInit {
   imageFile;
   @ViewChild("dealImage") dealImage: ElementRef;
   constructor(
-    private franchiseInfoService: FranchiseInfoService,
     private franchiseDealsService: FranchiseDealsService,
+    private isModal: IsModalService,
+    private toaster: IsToasterService,
     private storage: AngularFireStorage
   ) {}
 
@@ -58,19 +66,29 @@ export class DealsComponent implements OnInit {
     // this.showDeals = false;
     console.log("Edit Deal is : ", this.editDeal);
   }
-  onDeleteDealHandler(id) {
-    let delDeal = this.deals.filter(deal => deal.id == id);
-    this.deleteDeal = delDeal[0];
-    const delFile = this.storage.storage.refFromURL(this.deleteDeal.dealImage);
-    delFile.delete().then(deletedFile => {
-      this.franchiseDealsService.deleteDeal(id).subscribe(response => {
-        console.log("Response from Server : ", response);
-        this.deals = this.deals.filter(deal => deal.id != id);
-      });
+  onDeleteDealHandler(id, deleteDialog: TemplateRef<any>) {
+    const deleteModal = this.isModal.open(deleteDialog, {
+      data: "Are Your Sure you want to Delete this Deal ?"
+    });
+    deleteModal.onClose.subscribe(res => {
+      if (res === "ok") {
+        let delDeal = this.deals.filter(deal => deal.id == id);
+        this.deleteDeal = delDeal[0];
+        const delFile = this.storage.storage.refFromURL(
+          this.deleteDeal.dealImage
+        );
+        delFile.delete().then(deletedFile => {
+          this.franchiseDealsService.deleteDeal(id).subscribe(response => {
+            console.log("Response from Server : ", response);
+            this.toaster.popSuccess("Deal Has Been Deleted Successfully");
+            this.deals = this.deals.filter(deal => deal.id != id);
+          });
+        });
+      }
     });
   }
 
-  onDealSubmit(form: FormGroup, btn : IsButton) {
+  onDealSubmit(form: FormGroup, btn: IsButton) {
     btn.startLoading();
     let randomString =
       Math.random()
@@ -104,7 +122,7 @@ export class DealsComponent implements OnInit {
                   this.newDeal = responseData.data;
                   this.showDeals = true;
                   this.deals.push(this.newDeal);
-                  btn.stopLoading()
+                  btn.stopLoading();
                   console.log("this.newDeal : ", this.newDeal);
                   this.dealForm.reset();
                 });
